*** empty log message ***
svn path=/trunk/boinc/; revision=5477

+++ client/http.C
@@ -834,7 +834,7 @@ bool HTTP_OP_SET::poll(double) {
 							scope_messages.printf(
 								"HTTP_OP_SET::poll(): ERR_IO: bytes_xferred: %d,"
 								"file offset: %d, expected content length: %d\n",
-								htp->bytes_xferred, htp->file_offset, htp->hrh.content_length
+								(int)htp->bytes_xferred, (int)htp->file_offset, htp->hrh.content_length
 							);
                             htp->http_op_retval = ERR_IO;
                         }
